+++ crates/sswsort-cli/src/app/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod grid;
-pub(crate) mod write;
-
-pub(crate) use grid::*;
-pub(crate) use write::*;